Teamcenter aspectJ
This aspect fixes a performance problem in the Eclipse 3.8 mechanism
which is responsible for evaluating the visibleWhen and activeWhen etc
expressions defined in plugin.xml files that control the state of menu
items, commands or handlers. It implements a caching mechanism that
prevents that the same expressions are evaluated multiple times.